auth_service/requirements.txt
fastapi[all]==0.111.1
uvicorn==0.30.3
aiohttp==3.8.2
loguru==0.7.2
pyjwt==2.9.0